changed finalgame subcommand to CommandManager
